Доработал замечания
