+++ docs/com.example/Feline.java.html
@@ -13,7 +13,7 @@ import java.util.List;
     public String getFamily() {
 <span class="fc" id="L14">        return &quot;Кошачьи&quot;;</span>
     }
-
+    @Override
     public int getKittens() {
 <span class="fc" id="L18">        return getKittens(1);</span>
     }
