+++ src/main/java/com/example/Feline.java
@@ -13,7 +13,7 @@ public class Feline extends Animal implements Predator, IndependentLion {
     public String getFamily() {
         return "Кошачьи";
     }
-
+    @Override
     public int getKittens() {
         return getKittens(1);
     }
